requirements/common.txt
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==1.11.11
django-braces==1.13.0
django-oauth-toolkit==1.0.0
django-rest-framework-social-oauth2==1.0.8
djangorestframework==3.7.3
elasticsearch==6.3.1
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.1.0
olefile==0.45.1
Pillow==4.3.0
pkg-resources==0.0.0
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
uritemplate==3.0.0
urllib3==1.23
